pyproject.toml
[project]
name = "odm-kafka-bridge"
version = "0.1.0"
description = "A bridge that relays assets (e.g., DSMs) from WebODM to a Kafka cluster."
authors = [
	{ name="Merlin Stampa", email="merlin.stampa@rettungsrobotik.de" }
]
license = "GPL-3.0-or-later"
license-files = [ "LICENSE*" ]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
	"argparse",
	"confluent-kafka",
	"humanfriendly",
	"pathlib",
	"python-dotenv",
	"requests",
	"toml",
]

[dependency-groups]
dev = [
    "ruff",
]

[project.scripts]
odm-kafka-bridge = "odm_kafka_bridge.cli:main"

[project.urls]
Homepage = "https://crexdata.eu"
Issues = "https://github.com/rettungsrobotik/odm_kafka_bridge/issues"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.ruff.lint]
extend-select = ["I"]

[tool.setuptools]
package-dir = { "" = "src" }
packages = ["odm_kafka_bridge"]
